adding check operation to type system
