+++ bigraph_schema/type_system.py
@@ -8,18 +8,20 @@ import copy
 import pprint
 import pytest
 import random
+import numbers
 
+import numpy as np
 
 from bigraph_schema.react import react_divide_counts
-from bigraph_schema.base_types import base_type_library, set_apply, accumulate, concatenate, divide_float, divide_int, \
-    divide_longest, divide_list, replace, serialize_string, deserialize_string, to_string, deserialize_int, \
-    deserialize_float, evaluate, apply_any, serialize_any, deserialize_any, apply_tree, divide_tree, serialize_tree, deserialize_tree, apply_dict, divide_dict, \
-    serialize_dict, deserialize_dict, apply_maybe, divide_maybe, serialize_maybe, deserialize_maybe, apply_units, \
-    serialize_units, deserialize_units, divide_units, apply_edge, serialize_edge, deserialize_edge, divide_edge, \
-    serialize_list, deserialize_list, serialize_np_array, deserialize_np_array
+# from bigraph_schema.base_types import base_type_library, set_apply, accumulate, concatenate, divide_float, divide_int, \
+#     divide_longest, divide_list, replace, serialize_string, deserialize_string, to_string, deserialize_int, \
+#     deserialize_float, evaluate, apply_any, serialize_any, deserialize_any, apply_tree, divide_tree, serialize_tree, deserialize_tree, apply_dict, divide_dict, \
+#     serialize_dict, deserialize_dict, apply_maybe, divide_maybe, serialize_maybe, deserialize_maybe, apply_units, \
+#     serialize_units, deserialize_units, divide_units, apply_edge, serialize_edge, deserialize_edge, divide_edge, \
+#     serialize_list, deserialize_list, serialize_np_array, deserialize_np_array
 from bigraph_schema.registry import (
     Registry, TypeRegistry, RegistryRegistry, type_schema_keys, deep_merge, get_path,
-    establish_path, set_path, transform_path, remove_omitted, non_schema_keys
+    establish_path, set_path, transform_path, remove_path, remove_omitted, non_schema_keys
 )
 from bigraph_schema.units import units, render_units_type
 
@@ -36,12 +38,14 @@ class TypeSystem:
         self.serialize_registry = Registry(function_keys=['value', 'bindings', 'types'])
         self.deserialize_registry = Registry(function_keys=['serialized', 'bindings', 'types'])
         self.divide_registry = Registry()  # TODO enforce keys for divider methods
+        self.check_registry = Registry()
         self.react_registry = Registry()
         self.type_registry = TypeRegistry()
 
         self.registry_registry = RegistryRegistry()
         self.registry_registry.register('_type', self.type_registry)
         self.registry_registry.register('_react', self.react_registry)
+        self.registry_registry.register('_check', self.check_registry)
         self.registry_registry.register('_apply', self.apply_registry)
         self.registry_registry.register('_divide', self.divide_registry)
         self.registry_registry.register('_serialize', self.serialize_registry)
@@ -82,6 +86,7 @@ class TypeSystem:
 
     def access(self, type_key):
         return self.type_registry.access(type_key)
+
 
     def validate_schema(self, schema, enforce_connections=False):
         # add ports and wires
@@ -139,6 +144,7 @@ class TypeSystem:
         #             if not key in optional_schema_keys:
         #                 report[key] = f'missing required key: {key} for declaring atomic type'
 
+
     # TODO: if its an edge, ensure ports match wires
     def validate_state(self, original_schema, state):
         schema = self.access(original_schema)
@@ -172,6 +178,7 @@ class TypeSystem:
                             validation[key] = subvalidation
 
         return validation
+
 
     def default(self, schema):
         default = None
@@ -337,6 +344,45 @@ class TypeSystem:
 
         return state
 
+
+    def check_state(self, schema, state):
+        if isinstance(schema, str) or isinstance(schema, list):
+            # this assumes access always returns a dict
+            schema = self.access(schema)
+            return self.check_state(schema, state, update)
+
+        elif '_check' in schema and schema['_check'] != 'any':
+            check_function = self.check_registry.access(
+                schema['_check'])
+            
+            return check_function(
+                state,
+                schema.get('_bindings'),
+                self)
+
+        elif isinstance(state, dict):
+            for key, branch in schema.items():
+                if not key.startswith('_'):
+                    if key not in state:
+                        return False
+                    else:
+                        check = self.check_state(
+                            schema[key],
+                            state[key])
+
+                        if not check:
+                            return False
+
+        else:
+            return False
+
+        return True
+
+
+    def check(self, initial_schema, state):
+        schema = self.access(initial_schema)
+        return self.check_state(schema, state)
+    
 
     def apply_update(self, schema, state, update):
         if isinstance(update, dict) and '_react' in update:
@@ -895,6 +941,452 @@ class TypeSystem:
     #         return reactum
 
 
+NONE_SYMBOL = 'None'
+
+
+def instance_parameter(instance_type):
+    def is_instance_parameter(instance, bindings=None, types=None):
+        return isinstance(instance, instance_type)
+
+    return is_instance_parameter
+
+
+check_number = instance_parameter(numbers.Number)
+check_int = instance_parameter(int)
+check_float = instance_parameter(float)
+check_string = instance_parameter(str)
+
+
+def check_list(state, bindings, types):
+    element_type = bindings['element']
+    
+
+
+base_type_library = {
+    # abstract number type
+    'number': {
+        '_type': 'number',
+        '_apply': 'accumulate',
+        '_check': 'check_number',
+        '_serialize': 'to_string',
+        '_description': 'abstract base type for numbers'},
+
+    'int': {
+        '_type': 'int',
+        '_default': '0',
+        # inherit _apply and _serialize from number type
+        '_deserialize': 'deserialize_int',
+        '_divide': 'divide_int',
+        '_description': '64-bit integer',
+        '_super': 'number'},
+
+    'float': {
+        '_type': 'float',
+        '_default': '0.0',
+        '_deserialize': 'float',
+        '_divide': 'divide_float',
+        '_description': '64-bit floating point precision number',
+        '_super': 'number'},
+
+    'string': {
+        '_type': 'string',
+        '_default': '',
+        '_apply': 'replace',
+        '_serialize': 'serialize_string',
+        '_deserialize': 'deserialize_string',
+        '_divide': 'divide_int',
+        '_description': '64-bit integer'},
+
+    'list': {
+        '_type': 'list',
+        '_default': [],
+        '_apply': 'concatenate',
+        '_serialize': 'serialize_list',
+        '_deserialize': 'deserialize_list',
+        '_divide': 'divide_list',
+        '_type_parameters': ['element'],
+        '_description': 'general list type (or sublists)'},
+
+    'tree': {
+        '_type': 'tree',
+        '_default': {},
+        '_apply': 'apply_tree',
+        '_serialize': 'serialize_tree',
+        '_deserialize': 'deserialize_tree',
+        '_divide': 'divide_tree',
+        '_type_parameters': ['leaf'],
+        '_description': 'mapping from str to some type (or nested dicts)'},
+
+    'dict': {
+        '_type': 'dict',
+        '_default': {},
+        '_apply': 'apply_dict',
+        '_serialize': 'serialize_dict',
+        '_deserialize': 'deserialize_dict',
+        '_divide': 'divide_dict',
+        # TODO: create assignable type parameters?
+        '_type_parameters': ['key', 'value'],
+        '_description': 'mapping from keys of any type to values of any type'},
+
+    # TODO: add native numpy array type
+    'array': {
+        '_type': 'array',
+        '_type_parameters': ['shape', 'element']},
+
+    'maybe': {
+        '_type': 'maybe',
+        '_default': 'None',
+        '_apply': 'apply_maybe',
+        '_serialize': 'serialize_maybe',
+        '_deserialize': 'deserialize_maybe',
+        '_divide': 'divide_maybe',
+        '_type_parameters': ['value'],
+        '_description': 'type to represent values that could be empty'},
+
+    'edge': {
+        # TODO: do we need to have defaults informed by type parameters?
+        '_type': 'edge',
+        '_default': {'wires': {}},
+        '_apply': 'apply_edge',
+        '_serialize': 'serialize_edge',
+        '_deserialize': 'deserialize_edge',
+        '_divide': 'divide_edge',
+        '_type_parameters': ['ports'],
+        '_description': 'hyperedges in the bigraph, with ports as a type parameter',
+        'wires': 'tree[list[string]]',
+    },
+
+    'numpy_array': {
+        '_type': 'numpy_array',
+        '_default': np.array([]),
+        '_apply': 'accumulate',
+        '_serialize': 'serialize_np_array',
+        '_deserialize': 'deserialize_np_array',
+        '_description': 'numpy arrays'
+    },
+
+    'wires': 'tree[list[string]]',
+
+    # TODO -- this should support any type
+    'union': {
+        '_type': 'union',
+        '_default': 'None',
+        # '_apply': 'apply_maybe',
+        # '_serialize': 'serialize_maybe',
+        # '_deserialize': 'deserialize_maybe',
+        # '_divide': 'divide_maybe',
+        '_type_parameters': ['value'],
+        # '_description': 'type to represent values that could be empty'
+    },
+}
+
+
+#################
+# Apply methods #
+#################
+
+def apply_any(current, update, bindings=None, types=None):
+    return update
+
+
+def check_any(state, bindings=None, types=None):
+    return True
+
+
+def serialize_any(value, bindings=None, types=None):
+    return str(value)
+
+
+def accumulate(current, update, bindings=None, types=None):
+    if update is None:
+        return current
+    return current + update
+
+
+def set_apply(current, update, bindings=None, types=None):
+    return update
+
+
+def concatenate(current, update, bindings=None, types=None):
+    return current + update
+
+
+##################
+# Divide methods #
+##################
+# support dividing by ratios?
+# ---> divide_float({...}, [0.1, 0.3, 0.6])
+
+def divide_float(value, ratios, bindings=None, types=None):
+    half = value / 2.0
+    return (half, half)
+
+
+# support function types for registrys?
+# def divide_int(value: int, _) -> tuple[int, int]:
+def divide_int(value, bindings=None, types=None):
+    half = value // 2
+    other_half = half
+    if value % 2 == 1:
+        other_half += 1
+    return half, other_half
+
+
+def divide_longest(dimensions, bindings=None, types=None):
+    # any way to declare the required keys for this function in the registry?
+    # find a way to ask a function what type its domain and codomain are
+
+    width = dimensions['width']
+    height = dimensions['height']
+
+    if width > height:
+        a, b = divide_int(width)
+        return [{'width': a, 'height': height}, {'width': b, 'height': height}]
+    else:
+        x, y = divide_int(height)
+        return [{'width': width, 'height': x}, {'width': width, 'height': y}]
+
+
+def divide_list(l, bindings, types):
+    result = [[], []]
+    divide_type = bindings['element']
+    divide = divide_type['_divide']
+
+    for item in l:
+        if isinstance(item, list):
+            divisions = divide_list(item, bindings, types)
+        else:
+            divisions = divide(item, divide_type, types)
+
+        result[0].append(divisions[0])
+        result[1].append(divisions[1])
+
+    return result
+
+
+def replace(current, update, bindings=None, types=None):
+    return update
+
+
+#####################
+# Serialize methods #
+#####################
+
+def serialize_string(value, bindings=None, types=None):
+    return value
+
+
+def deserialize_string(serialized, bindings=None, types=None):
+    return serialized
+
+
+def to_string(value, bindings=None, types=None):
+    return str(value)
+
+
+def serialize_list(value, bindings=None, types=None):
+    schema = bindings['element']
+    return [types.serialize(schema, element) for element in value]
+
+def serialize_np_array(value, bindings=None, types=None):
+    ''' Serialize numpy array to bytes '''
+    return {
+        'bytes': value.tobytes(),
+        'dtype': value.dtype,
+        'shape': value.shape,
+    }
+
+
+#######################
+# Deserialize methods #
+#######################
+
+def deserialize_int(serialized, bindings=None, types=None):
+    return int(serialized)
+
+
+def deserialize_float(serialized, bindings=None, types=None):
+    return float(serialized)
+
+
+def evaluate(serialized, bindings=None, types=None):
+    return eval(serialized)
+
+
+def deserialize_any(serialized, bindings=None, types=None):
+    return serialized
+
+def deserialize_list(serialized, bindings=None, types=None):
+    schema = bindings['element']
+    return [types.deserialize(schema, element) for element in serialized]
+
+
+def deserialize_np_array(serialized, bindings=None, types=None):
+    if isinstance(serialized, dict):
+        np.frombuffer(serialized['bytes'], dtype=serialized['dtype']).reshape(serialized['shape'])
+        return np.frombuffer(serialized)
+    else:
+        return  serialized
+
+
+# TODO: make these work
+def apply_tree(current, update, bindings, types):
+    leaf_type = types.access(bindings['leaf'])
+    bindings['leaf'] = leaf_type
+    
+    if isinstance(update, dict):
+        current = current or {}
+        
+        for key, branch in update.items():
+            if key == '_add':
+                current.update(branch)
+            elif key == '_remove':
+                current = remove_path(current, branch)
+            else:
+                current[key] = apply_tree(
+                    current.get(key),
+                    branch,
+                    bindings,
+                    types)
+
+        return current
+    else:
+        if current is None:
+            current = types.default(leaf_type)
+        return types.apply(leaf_type, current, update)
+
+
+def check_tree(tree, bindings, types):
+    leaf_type = bindings['leaf']
+    return types.check(leaf_type, tree)
+
+    # for key, value in tree.items():
+    #     check = types.check(leaf_type, value)
+    #     if not check:
+    #         return False
+
+
+def divide_tree(tree, bindings, types):
+    result = [{}, {}]
+    # get the type of the values for this dict
+    divide_type = bindings['leaf']
+    divide_function = divide_type['_divide']
+    # divide_function = types.registry_registry.type_attribute(
+    #     divide_type,
+    #     '_divide')
+
+    for key, value in tree.items():
+        if isinstance(value, dict):
+            divisions = divide_tree(value)
+        else:
+            divisions = types.divide(divide_type, value)
+
+        result[0][key], result[1][key] = divisions
+
+    return result
+
+
+def serialize_tree(value, bindings=None, types=None):
+    return value
+
+
+def deserialize_tree(serialized, bindings=None, types=None):
+    tree = serialized
+
+    if isinstance(serialized, str):
+        tree = types.deserialize(
+            bindings['leaf'],
+            serialized)
+
+    elif isinstance(serialized, dict):
+        tree = {}
+        for key, value in serialized.items():
+            tree[key] = deserialize_tree(value, bindings, types)
+
+    return tree
+
+
+def apply_dict(current, update, bindings=None, types=None):
+    pass
+
+
+def divide_dict(value, bindings=None, types=None):
+    return value
+
+
+def serialize_dict(value, bindings=None, types=None):
+    return value
+
+
+def deserialize_dict(serialized, bindings=None, types=None):
+    return serialized
+
+
+def apply_maybe(current, update, bindings, types):
+    if current is None or update is None:
+        return update
+    else:
+        value_type = bindings['value']
+        return types.apply(value_type, current, update)
+
+
+def divide_maybe(value, bindings):
+    if value is None:
+        return [None, None]
+    else:
+        pass
+
+
+def serialize_maybe(value, bindings, types):
+    if value is None:
+        return NONE_SYMBOL
+    else:
+        value_type = bindings['value']
+        return serialize(value_type, value)
+
+
+def deserialize_maybe(serialized, bindings, types):
+    if serialized == NONE_SYMBOL:
+        return None
+    else:
+        value_type = bindings['value']
+        return types.deserialize(value_type, serialized)
+
+
+# TODO: deal with all the different unit types
+def apply_units(current, update, bindings, types):
+    return current + update
+
+
+def serialize_units(value, bindings, types):
+    return str(value)
+
+
+def deserialize_units(serialized, bindings, types):
+    return units(serialized)
+
+
+def divide_units(value, bindings, types):
+    return [value, value]
+
+
+# TODO: implement edge handling
+def apply_edge(current, update, bindings, types):
+    return current + update
+
+
+def serialize_edge(value, bindings, types):
+    return str(value)
+
+
+def deserialize_edge(serialized, bindings, types):
+    return eval(serialized)
+
+
+def divide_edge(value, bindings, types):
+    return [value, value]
+
+
 def register_units(types, units):
     for unit_name in units._units:
         try:
@@ -938,6 +1430,16 @@ def register_base_types(types):
     types.divide_registry.register('divide_maybe', divide_maybe)
     types.divide_registry.register('divide_units', divide_units)
     types.divide_registry.register('divide_edge', divide_edge)
+
+    types.check_registry.register('check_float', check_float)
+    types.check_registry.register('check_int', check_int)
+    types.check_registry.register('check_longest', check_longest)
+    types.check_registry.register('check_list', check_list)
+    types.check_registry.register('check_tree', check_tree)
+    types.check_registry.register('check_dict', check_dict)
+    types.check_registry.register('check_maybe', check_maybe)
+    types.check_registry.register('check_units', check_units)
+    types.check_registry.register('check_edge', check_edge)
 
     types.serialize_registry.register('serialize_any', serialize_any)
     types.serialize_registry.register('serialize_string', serialize_string)
